Separating form from create user form so that it can be reused for login
